+++ src/main/java/local/springdemobot/database/repository/UserStatusRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserStatusRepository extends JpaRepository<UserStatus, Integer> {
+public interface UserStatusRepository extends JpaRepository<UserStatus, Long> {
 }
